+++ Peach.Core.Analysis.Pin.CoverageEdge/cedge.cpp
@@ -115,7 +115,7 @@ VOID Fini(INT32 code, VOID *v)
 	{
 		for (set<ADDRINT>::iterator i = setKnownBlocks.begin(); i!=setKnownBlocks.end(); ++i)
 		{
-			fprintf(existing, "%16.16zx\n", *i);
+			fprintf(existing, "%16.16zx\n", (size_t)*i);
 		}
 
 		fclose(existing);
@@ -127,7 +127,7 @@ VOID Fini(INT32 code, VOID *v)
 	{
 		for (set<ADDRINT>::iterator i = setUnknownBlocks.begin(); i != setUnknownBlocks.end(); ++i)
 		{
-			fprintf(existing, "%16.16zx\n", *i);
+			fprintf(existing, "%16.16zx\n", (size_t)*i);
 		}
 
 		fclose(existing);
